fix var size..

+++ DAQBoard_comm.h
@@ -107,7 +107,7 @@ private:
 		bool spi_unavaiable;
 
 		std::array<uint16_t, GCR_address_max> GCR_address_array;
-		std::array<uint16_t, ctrl_address_max> ctrl_address_array;
+		std::array<uint32_t, ctrl_address_max> ctrl_address_array;
 
 		chip_struct() : run_flag({false}), spi_unavaiable(false), ctrl_address_array({0}) {}
 	};
